refactor(personhog): move the merge entrance out of the lifecycle service
The lifecycle service was accumulating identity work: MergePersons'
validation, distinct id resolution, classification, and inline
settlement, plus the storage and leader handles they need. That work is
identity's, not the saga machinery's, so it moves to
service::merge::MergeEntrance. The lifecycle side exposes exactly one
seam — MergeOpExecutor, attach to an existing op row or drive a frozen
request to its terminal row — and the service's merge_persons handler
becomes a one-line delegation. If the saga RPCs later move to an
internal service, the executor is already the boundary a wire would
replace.

No behavior change; the entrance body is the previous handler verbatim.
