+++ rust/personhog-identity/src/main.rs
@@ -23,8 +23,9 @@ use personhog_identity::config::Config;
 use personhog_identity::leader::LifecycleLeader;
 use personhog_identity::lifecycle::delete::DeleteDriver;
 use personhog_identity::lifecycle::engine::Engine;
-use personhog_identity::lifecycle::merge::MergeDriver;
+use personhog_identity::lifecycle::merge::{MergeDriver, MergeOpExecutor};
 use personhog_identity::lifecycle::PersonHogLifecycleService;
+use personhog_identity::service::merge::MergeEntrance;
 use personhog_identity::service::PersonHogIdentityService;
 use personhog_identity::storage::postgres::PostgresIdentityStorage;
 
@@ -235,14 +236,16 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     // Separate proto service co-served on the same server so lifecycle
     // callers are insulated from any future split.
-    let lifecycle_service = PersonHogLifecycleService::new(
-        engine,
-        lifecycle_leader,
-        config.tables(),
+    let merge_entrance = MergeEntrance::new(
         storage.clone(),
         property_writer.clone(),
-        MergeDriver::new(property_writer.clone(), config.tables()),
+        MergeOpExecutor::new(
+            engine.clone(),
+            MergeDriver::new(property_writer.clone(), config.tables()),
+        ),
     );
+    let lifecycle_service =
+        PersonHogLifecycleService::new(engine, lifecycle_leader, config.tables(), merge_entrance);
     let service = PersonHogIdentityService::new(storage, property_writer, config.request_limits());
 
     let grpc_addr = config.grpc_address;
